+++ Gems/EMotionFX/Code/EMotionFX/Source/SubMesh.h
@@ -55,7 +55,7 @@ namespace EMotionFX
          * @param materialIndex The material.
          * @param numBones The number of bones inside the submesh.
          */
-        static SubMesh* Create(Mesh* parentMesh, uint32 startVertex, uint32 startIndex, uint32 startPolygon, uint32 numVerts, uint32 numIndices, uint32 numPolygons, uint32 materialIndex, uint32 numBones);
+        static SubMesh* Create(Mesh* parentMesh, uint32 startVertex, uint32 startIndex, uint32 startPolygon, uint32 numVerts, uint32 numIndices, uint32 numPolygons, uint32 materialIndex, size_t numBones);
 
         /**
          * Get the start index. This is the offset in the index array of the parent mesh where the index data for this
@@ -178,14 +178,14 @@ namespace EMotionFX
          * Set the number of bones that is being used by this submesh.
          * @param numBones The number of bones used by the submesh.
          */
-        void SetNumBones(uint32 numBones);
+        void SetNumBones(size_t numBones);
 
         /**
          * Set the index of a given bone.
          * @param index The bone number, which must be in range of [0..GetNumBones()-1].
          * @param nodeIndex The node index number that acts as bone on this submesh.
          */
-        void SetBone(uint32 index, uint32 nodeIndex);
+        void SetBone(size_t index, size_t nodeIndex);
 
         /**
          * Get the number of bones used by this submesh.
@@ -236,20 +236,20 @@ namespace EMotionFX
          * @result The bone number inside the submesh, which is in range of [0..GetNumBones()-1].
          *         A value of MCORE_INVALIDINDEX32 is returned when the specified node isn't used as bone inside this submesh.
          */
-        uint32 FindBoneIndex(uint32 nodeNr) const;
+        size_t FindBoneIndex(size_t nodeNr) const;
 
         /**
          * Remap bone to a new bone. This will overwrite the given old bones with the new one.
          * @param oldNodeNr The node number to be searched and replaced.
          * @param newNodeNr The node number with which the old bones will be replaced with.
          */
-        void RemapBone(uint16 oldNodeNr, uint16 newNodeNr);
+        void RemapBone(size_t oldNodeNr, size_t newNodeNr);
 
         /**
          * Remove the given bone from the bones list.
          * @param index The index of the bone to be removed in range of [0..GetNumBones()-1].
          */
-        void RemoveBone(uint16 index);
+        void RemoveBone(size_t index);
 
         /**
          * Clone the submesh.
@@ -290,7 +290,7 @@ namespace EMotionFX
          * @param materialIndex The material.
          * @param numBones The number of bones inside the submesh.
          */
-        SubMesh(Mesh* parentMesh, uint32 startVertex, uint32 startIndex, uint32 startPolygon, uint32 numVerts, uint32 numIndices, uint32 numPolygons, uint32 materialIndex, uint32 numBones);
+        SubMesh(Mesh* parentMesh, uint32 startVertex, uint32 startIndex, uint32 startPolygon, uint32 numVerts, uint32 numIndices, uint32 numPolygons, uint32 materialIndex, size_t numBones);
 
         /**
          * Destructor.
